put pad-utils back in pad
