+++ examples/notepad/index.html
@@ -118,15 +118,16 @@
   <h2>manageParticipation</h2>
   <script id='script-manageParticipation'>
     window.addEventListener('DOMContentLoaded', async (event) => {
-      const doc = $('#frame-manageParticipation').contentWindow.document;
+      const doc = $('#iframe-manageParticipation').contentWindow.document;
       const kb = UI.store
       const me = new UI.rdf.NamedNode('https://sharonstrats.inrupt.net/profile/card#me')
-      const container = createElement(HTMLDivElement)
+      const container = document.createElement('div')
       const subject = kb.sym('https://test.test#');
+      const fetcher = UI.rdf.fetcher(kb);
       const padDoc = kb.sym(document.location.href.replace(/\/[^\/]*$/, "/pad.ttl#this"))
       await fetcher.load(padDoc.doc())
       options = {}
-      const htmlStr = manageParticipation(Document, container, padDoc, subject, me, options)
+      const htmlStr = UI.pad.manageParticipation(doc, container, padDoc, subject, me, options)
     })
   </script>
 
@@ -137,15 +138,16 @@
   <h2>notepad </h2>
   <script id='script-notepad'>
     window.addEventListener('DOMContentLoaded', async (event) => {
-      const doc = $('#frame-notepad').contentWindow.document;
+      const doc = $('#iframe-notepad').contentWindow.document;
       const kb = UI.store
       const me = new UI.rdf.NamedNode('https://sharonstrats.inrupt.net/profile/card#me')
-      const container = createElement(HTMLDivElement)
+      const container = document.createElement('div')
       const subject = kb.sym('https://test.test#');
+      const fetcher = UI.rdf.fetcher(kb);
       const padDoc = kb.sym(document.location.href.replace(/\/[^\/]*$/, "/pad.ttl#this"))
       await fetcher.load(padDoc.doc())
       options = {}
-      const htmlStr = notepad(Document, padDoc, subject, me, options)
+      const htmlStr = UI.pad.notepad(doc, padDoc, subject, me, options)
     })
   </script>
   <pre id="viewSource-notepad"></pre>
